+++ ui/CHANGELOG.md
@@ -2,9 +2,19 @@
 
 ## Unreleased
 
+## [0.111.1] - 2023-11-21
+
+### Fixed
+- Filter simple condition: display all variables when using a date column
+
+## [0.111.0] - 2023-10-06
+
+### Added
+- Waterfall step: Added support for the `backfill` option
+
 ## [0.110.0] - 2023-09-26
 
-## Removed
+### Removed
 - Remove `RELATIVE_DATE_FILTERING` feature flag and related code
 
 ## [0.109.0] - 2023-09-08
@@ -1531,6 +1541,8 @@ TabbedRangeCalendars: keep selected tab unchanged when updated enabled calendars
 
 - Initial version, showtime!
 
+[0.111.1]: https://github.com/ToucanToco/weaverbird/compare/v0.111.0...v0.111.1
+[0.111.0]: https://github.com/ToucanToco/weaverbird/compare/v0.110.0...v0.111.0
 [0.110.0]: https://github.com/ToucanToco/weaverbird/compare/v0.109.0...v0.110.0
 [0.109.0]: https://github.com/ToucanToco/weaverbird/compare/v0.108.0...v0.109.0
 [0.108.0]: https://github.com/ToucanToco/weaverbird/compare/v0.107.0...v0.108.0